Create Authors
As the contributors on several of the files were deleted, for the purposes of correctly assigning pull requests, I have trawled through the history and attempted to discover who first authored each readme. As you can see, I failed to find the author in a couple of places and I probably made some mistakes in others, so it would be great if anyone who sees this can check that their own attribution is correct and amend as necessary. Thanks!